+++ src/Attribute/MapTo.php
@@ -11,13 +11,14 @@ namespace AutoMapper\Attribute;
 final readonly class MapTo
 {
     /**
-     * @param class-string<object>|'array'|array<class-string<object>|'array'>|null $target      The specific target class name or array. If null this attribute will be used for all target classes.
-     * @param string|null                                                           $property    The target property name. If null, the source property name will be used.
-     * @param int|null                                                              $maxDepth    The maximum depth of the mapping. If null, the default max depth will be used.
-     * @param string|callable(mixed $value, object $object): mixed|null             $transformer A transformer id or a callable that transform the value during mapping
-     * @param bool|null                                                             $ignore      if true, the property will be ignored during mapping
-     * @param string|null                                                           $if          The condition to map the property, using the expression language
-     * @param string[]|null                                                         $groups      The groups to map the property
+     * @param class-string<object>|'array'|array<class-string<object>|'array'>|null $target         The specific target class name or array. If null this attribute will be used for all target classes.
+     * @param string|null                                                           $property       The target property name. If null, the source property name will be used.
+     * @param int|null                                                              $maxDepth       The maximum depth of the mapping. If null, the default max depth will be used.
+     * @param string|callable(mixed $value, object $object): mixed|null             $transformer    A transformer id or a callable that transform the value during mapping
+     * @param bool|null                                                             $ignore         If true, the property will be ignored during mapping
+     * @param string|null                                                           $if             The condition to map the property, using the expression language
+     * @param string[]|null                                                         $groups         The groups to map the property
+     * @param string|null                                                           $dateTimeFormat The date-time format to use when transforming this property
      */
     public function __construct(
         public string|array|null $target = null,
@@ -28,6 +29,7 @@ final readonly class MapTo
         public ?string $if = null,
         public ?array $groups = null,
         public int $priority = 0,
+        public ?string $dateTimeFormat = null,
     ) {
     }
 }